Simplify default layout template by moving banners to components
