+++ apps/frontend/src/composables/nuxt-accessors.ts
@@ -1,7 +1,9 @@
+import type { Cosmetics } from '~/plugins/cosmetics.ts'
+
 export function useTheme() {
 	return useNuxtApp().$theme
 }
 
 export function useCosmetics() {
-	return useNuxtApp().$cosmetics
+	return useNuxtApp().$cosmetics as Ref<Cosmetics>
 }
